Backend/requirements.txt
flask
flask-cors
numpy
pandas
scikit-learn
joblib
xgboost
gunicorn
mysql-connector-python
